removed middleware..added to users/[id]
