+++ lib/server/firebaseAdmin.ts
@@ -1,0 +1,27 @@
+// @lib/server/firebaseAdmin.ts
+import { initializeApp, getApps, cert } from 'firebase-admin/app';
+import { getAuth } from 'firebase-admin/auth';
+
+// Initialize Firebase Admin SDK if not already initialized
+export const initializeFirebaseAdmin = () => {
+  if (getApps().length === 0) {
+    // Check for environment variables
+    const privateKey = process.env.FIREBASE_PRIVATE_KEY
+      ? process.env.FIREBASE_PRIVATE_KEY.replace(/\\n/g, '\n')
+      : undefined;
+
+    if (!privateKey || !process.env.FIREBASE_CLIENT_EMAIL || !process.env.FIREBASE_PROJECT_ID) {
+      console.error('Firebase Admin SDK credentials missing');
+      throw new Error('Firebase Admin credentials not found');
+    }
+
+    initializeApp({
+      credential: cert({
+        projectId: process.env.FIREBASE_PROJECT_ID,
+        clientEmail: process.env.FIREBASE_CLIENT_EMAIL,
+        privateKey: privateKey,
+      }),
+    });
+  }
+  return getAuth();
+};
